backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.2
werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.1

psycopg2-binary==2.9.9
sqlalchemy==2.0.21

cryptography==42.0.5
email-validator==2.1.0

python-dateutil==2.8.2
requests==2.31.0

numpy==2.4.2
pandas==2.2.2
scikit-learn==1.5.2
joblib==1.4.2

matplotlib==3.8.4
Pillow==10.2.0
